Adds compiler warnings to initial repo configuration

+++ src/sensor_fusion/include/env_state.h
@@ -29,7 +29,7 @@ class EnvironmentState {
   common::tracked_output_msg get_tracked_output_msg();
 
   void add_object(const ObjectState& tracked_msg);
-  void update_object(const ObjectState& tracked_msg, int index);
+  void update_object(const ObjectState& tracked_msg, size_t index);
   void check_timestamp(const ObjectState& tracked_msg);
   void update_env_state(const ObjectState& tracked_msg); 
   void find_target_objects(const ObjectState& tracked_msg);
